refactor(ui-study): clean up NavHeader component for better maintainability
