+++ webapp/public/locales/en/main.json
@@ -574,8 +574,8 @@
   "studies.enterTag": "Enter a tag",
   "studies.favorites": "Favorites",
   "studies.tagsLabel": "Tags",
-  "studies.bookmark": "Bookmark",
   "studies.moreActions": "More actions",
+  "studies.addFavorite": "Add to favorites",
   "studies.removeFavorite": "Remove from favorites",
   "studies.exportOptions": "Export options",
   "studies.exportOutput": "Export particular output",
