+++ webapp/public/locales/fr/main.json
@@ -574,8 +574,8 @@
   "studies.enterTag": "Entrer un tag",
   "studies.favorites": "Favoris",
   "studies.tagsLabel": "Tags",
-  "studies.bookmark": "Bookmark",
   "studies.moreActions": "Plus d'actions",
+  "studies.addFavorite": "Ajouter aux favoris",
   "studies.removeFavorite": "Retirer des favoris",
   "studies.exportOptions": "Options d'export",
   "studies.exportOutput": "Exporter une sortie",
