+++ webapp/src/components/App/Singlestudy/NavHeader/index.tsx
@@ -1,0 +1,343 @@
+import { useState, MouseEvent } from "react";
+import debug from "debug";
+import { useSnackbar } from "notistack";
+import { useNavigate } from "react-router-dom";
+import { AxiosError } from "axios";
+import { Box, Button, Tooltip, Typography } from "@mui/material";
+import ArrowBackIcon from "@mui/icons-material/ArrowBack";
+import MoreVertIcon from "@mui/icons-material/MoreVert";
+import ErrorIcon from "@mui/icons-material/Error";
+import UnarchiveOutlinedIcon from "@mui/icons-material/UnarchiveOutlined";
+import DownloadOutlinedIcon from "@mui/icons-material/DownloadOutlined";
+import ArchiveOutlinedIcon from "@mui/icons-material/ArchiveOutlined";
+import DeleteOutlinedIcon from "@mui/icons-material/DeleteOutlined";
+import EditOutlinedIcon from "@mui/icons-material/EditOutlined";
+import { useTranslation } from "react-i18next";
+import { StudyMetadata, VariantTree } from "../../../../common/types";
+import { STUDIES_HEIGHT_HEADER } from "../../../../theme";
+import { archiveStudy, unarchiveStudy } from "../../../../services/api/study";
+import { deleteStudy } from "../../../../redux/ducks/studies";
+import LauncherDialog from "../../Studies/LauncherDialog";
+import PropertiesDialog from "../PropertiesDialog";
+import useEnqueueErrorSnackbar from "../../../../hooks/useEnqueueErrorSnackbar";
+import { getLatestStudyVersion } from "../../../../redux/selectors";
+import ExportDialog from "../../Studies/ExportModal";
+import ConfirmationDialog from "../../../common/dialogs/ConfirmationDialog";
+import useAppSelector from "../../../../redux/hooks/useAppSelector";
+import useAppDispatch from "../../../../redux/hooks/useAppDispatch";
+import CheckBoxFE from "../../../common/fieldEditors/CheckBoxFE";
+import NavHeaderInfo from "./NavHeaderInfo";
+import NavHeaderCommands from "./NavHeaderCommands";
+import UpgradeDialog from "../UpgradeDialog";
+import NavHeaderMenu, { NavMenuItem } from "./NavHeaderMenu";
+
+const logError = debug("antares:singlestudy:navheader:error");
+
+interface Props {
+  study: StudyMetadata | undefined;
+  parent: StudyMetadata | undefined;
+  childrenTree: VariantTree | undefined;
+  isExplorer?: boolean;
+  openCommands: VoidFunction;
+  updateStudyData: VoidFunction;
+}
+
+function NavHeader({
+  study,
+  parent,
+  childrenTree,
+  isExplorer,
+  openCommands,
+  updateStudyData,
+}: Props) {
+  const [t] = useTranslation();
+  const navigate = useNavigate();
+  const dispatch = useAppDispatch();
+  const latestVersion = useAppSelector(getLatestStudyVersion);
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+  const [openMenu, setOpenMenu] = useState("");
+  const [openLauncherDialog, setOpenLauncherDialog] = useState(false);
+  const [openPropertiesDialog, setOpenPropertiesDialog] = useState(false);
+  const [openUpgradeDialog, setOpenUpgradeDialog] = useState(false);
+  const [deleteChildren, setDeleteChildren] = useState(false);
+  const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
+  const [openExportDialog, setOpenExportDialog] = useState(false);
+  const { enqueueSnackbar } = useSnackbar();
+  const enqueueErrorSnackbar = useEnqueueErrorSnackbar();
+  const isLatestVersion = study?.version === latestVersion;
+  const isManaged = study?.managed;
+  const isArchived = study?.archived;
+
+  ////////////////////////////////////////////////////////////////
+  // Event Handlers
+  ////////////////////////////////////////////////////////////////
+
+  const handleClick = (event: MouseEvent<HTMLButtonElement>) => {
+    setAnchorEl(event.currentTarget);
+    setOpenMenu(event.currentTarget.id);
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+    setOpenMenu("");
+  };
+
+  const handleClickBack = () => {
+    if (isExplorer) {
+      navigate(`/studies/${study?.id}`);
+    } else {
+      navigate("/studies");
+    }
+  };
+
+  const handleLaunch = () => {
+    setOpenLauncherDialog(true);
+  };
+
+  const handleArchive = async () => {
+    if (study) {
+      try {
+        await archiveStudy(study.id);
+      } catch (e) {
+        enqueueErrorSnackbar(
+          t("studies.error.archive", { studyname: study.name }),
+          e as AxiosError
+        );
+      } finally {
+        handleClose();
+      }
+    }
+  };
+
+  const handleUnarchive = async () => {
+    if (study) {
+      try {
+        await unarchiveStudy(study.id);
+      } catch (e) {
+        enqueueErrorSnackbar(
+          t("studies.error.unarchive", { studyname: study.name }),
+          e as AxiosError
+        );
+      } finally {
+        handleClose();
+      }
+    }
+  };
+
+  const handleDelete = async () => {
+    if (study) {
+      try {
+        await dispatch(deleteStudy({ id: study.id, deleteChildren })).unwrap();
+        navigate(parent ? `/studies/${parent?.id}` : "/studies");
+      } catch (err) {
+        enqueueErrorSnackbar(t("studies.error.deleteStudy"), err as AxiosError);
+        logError("Failed to delete study", study, err);
+      } finally {
+        setDeleteChildren(false);
+        setOpenDeleteDialog(false);
+      }
+    }
+  };
+
+  const handleCopyId = async () => {
+    if (study) {
+      try {
+        await navigator.clipboard.writeText(study.id);
+        enqueueSnackbar(t("study.success.studyIdCopy"), {
+          variant: "success",
+        });
+      } catch (e) {
+        enqueueErrorSnackbar(t("study.error.studyIdCopy"), e as AxiosError);
+      }
+    }
+  };
+
+  ////////////////////////////////////////////////////////////////
+  // Utils
+  ////////////////////////////////////////////////////////////////
+
+  const menuItems: NavMenuItem[] = [
+    {
+      key: "study.properties",
+      icon: EditOutlinedIcon,
+      action: () => setOpenPropertiesDialog(true),
+      condition: !isArchived,
+    },
+    {
+      key: "study.upgrade",
+      icon: ErrorIcon,
+      action: () => setOpenUpgradeDialog(true),
+      condition: !isArchived && !isLatestVersion,
+    },
+    {
+      key: "global.export",
+      icon: DownloadOutlinedIcon,
+      action: () => setOpenExportDialog(true),
+      condition: !isArchived,
+    },
+    {
+      key: "global.archive",
+      icon: ArchiveOutlinedIcon,
+      action: handleArchive,
+      condition: !isArchived && isManaged,
+    },
+    {
+      key: "global.unarchive",
+      icon: UnarchiveOutlinedIcon,
+      action: handleUnarchive,
+      condition: isArchived,
+    },
+    {
+      key: "global.delete",
+      icon: DeleteOutlinedIcon,
+      action: () => setOpenDeleteDialog(true),
+      color: "error.light",
+      condition: isManaged,
+    },
+  ];
+
+  ////////////////////////////////////////////////////////////////
+  // JSX
+  ////////////////////////////////////////////////////////////////
+
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "flex-start",
+        alignItems: "center",
+        width: 1,
+        height: STUDIES_HEIGHT_HEADER,
+        p: 2,
+        overflow: "hidden",
+      }}
+    >
+      <Box
+        sx={{
+          width: 1,
+          display: "flex",
+          flexDirection: "row",
+          justifyContent: "flex-start",
+          alignItems: "center",
+        }}
+      >
+        <ArrowBackIcon
+          color="secondary"
+          onClick={handleClickBack}
+          sx={{ cursor: "pointer" }}
+        />
+        <Button variant="text" color="secondary" onClick={handleClickBack}>
+          <Tooltip
+            title={isExplorer ? study?.name : t("global.studies")}
+            followCursor
+          >
+            <Typography variant="button">
+              {isExplorer ? t("button.back") : t("global.studies")}
+            </Typography>
+          </Tooltip>
+        </Button>
+      </Box>
+      <Box
+        sx={{
+          width: 1,
+          display: "flex",
+        }}
+      >
+        <NavHeaderCommands
+          study={study}
+          onOpenCommands={openCommands}
+          isExplorer={isExplorer}
+          onCopyId={handleCopyId}
+          onLaunch={handleLaunch}
+          onUnarchive={handleUnarchive}
+        />
+        <Button
+          size="small"
+          aria-controls="menu-study"
+          aria-haspopup="true"
+          id="menu-study"
+          variant="outlined"
+          color="primary"
+          sx={{ width: "auto", minWidth: 0, px: 0, marginLeft: 2 }}
+          onClick={handleClick}
+        >
+          <MoreVertIcon />
+        </Button>
+        <NavHeaderMenu
+          anchorEl={anchorEl}
+          openMenu={openMenu}
+          onClose={handleClose}
+          menuItems={menuItems}
+        />
+      </Box>
+      <NavHeaderInfo
+        study={study}
+        parent={parent}
+        childrenTree={childrenTree}
+      />
+      {study && openLauncherDialog && (
+        <LauncherDialog
+          open={openLauncherDialog}
+          studyIds={[study.id]}
+          onClose={() => setOpenLauncherDialog(false)}
+        />
+      )}
+      {study && openPropertiesDialog && (
+        <PropertiesDialog
+          open={openPropertiesDialog}
+          onClose={() => setOpenPropertiesDialog(false)}
+          study={study}
+          updateStudyData={updateStudyData}
+        />
+      )}
+      {openDeleteDialog && (
+        <ConfirmationDialog
+          title={t("dialog.title.confirmation")}
+          onCancel={() => setOpenDeleteDialog(false)}
+          onConfirm={handleDelete}
+          alert="warning"
+          open
+        >
+          <Box
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "flex-start",
+              width: 1,
+              height: "auto",
+              gap: 1,
+            }}
+          >
+            <Typography>{t("studies.question.delete")}</Typography>
+            <CheckBoxFE
+              value={deleteChildren}
+              label={t("studies.deleteSubvariants")}
+              onChange={(_e, checked) => setDeleteChildren(checked)}
+            />
+          </Box>
+        </ConfirmationDialog>
+      )}
+      {openUpgradeDialog && (
+        <ConfirmationDialog
+          title={t("dialog.title.confirmation")}
+          onCancel={() => setOpenUpgradeDialog(false)}
+          onConfirm={handleUpgrade}
+          alert="warning"
+          open
+        >
+          <Typography>{t("study.question.upgrade")}</Typography>
+        </ConfirmationDialog>
+      )}
+      {study && openExportDialog && (
+        <ExportDialog
+          open={openExportDialog}
+          onClose={() => setOpenExportDialog(false)}
+          study={study}
+        />
+      )}
+    </Box>
+  );
+}
+
+export default NavHeader;
